+++ weblate/trans/autofixes/chars.py
@@ -6,10 +6,7 @@ import re
 
 from django.utils.translation import gettext_lazy
 
-from weblate.checks.chars import (
-    FRENCH_PUNCTUATION_FIXUP_RE,
-    FRENCH_PUNCTUATION_MISSING_RE,
-)
+from weblate.checks.chars import FRENCH_PUNCTUATION_FIXUP_RE
 from weblate.formats.helpers import CONTROLCHARS_TRANS
 from weblate.trans.autofixes.base import AutoFix
 
@@ -81,9 +78,7 @@ class PunctuationSpacing(AutoFix):
         ):
             # Fix existing
             new_target = re.sub(FRENCH_PUNCTUATION_FIXUP_RE, "\u202F\\2", target)
-            # Add missing
-            new_target = re.sub(
-                FRENCH_PUNCTUATION_MISSING_RE, "\\1\u202F\\2", new_target
-            )
+            # Do not add missing as that is likely to trigger issues with other content
+            # such as URLs or Markdown syntax.
             return new_target, new_target != target
         return target, False